cdc/requirements.txt
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-bigquery==3.34.0
google-cloud-bigquery-storage==2.32.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
grpcio-tools==1.73.0
idna==3.10
loguru==0.7.3
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
wheel==0.45.1
